[OTHER] not.empy.annotation-update
Firtly, I added to all classes that are not subtypes, a @JsonTypeInfo annotation for proper class serialiazation of information in jsons. Secondly, added a @NotEmpty annotations that throw exceptions when a string is empty or if the field is not even a String. We can modify it to work with other types later, such as lists.